pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.28.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.28",
    "mss>=10.0.0",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyaudio>=0.2.14",
    "uvicorn>=0.35.0",
]
